Implement `and` and `or` in `Matcher` directly
