+++ core/src/com/google/inject/matcher/Matcher.java
@@ -30,11 +30,15 @@ public interface Matcher<T> {
    * Returns a new matcher which returns {@code true} if both this and the given matcher return
    * {@code true}.
    */
-  Matcher<T> and(Matcher<? super T> other);
+  default Matcher<T> and(Matcher<? super T> other) {
+    return new Matchers.AndMatcher<T>(this, other);
+  }
 
   /**
    * Returns a new matcher which returns {@code true} if either this or the given matcher return
    * {@code true}.
    */
-  Matcher<T> or(Matcher<? super T> other);
+  default Matcher<T> or(Matcher<? super T> other) {
+    return new Matchers.OrMatcher<T>(this, other);
+  }
 }
